Refactor Phase 4: extract WordConverter/LatexConverter, remove inspect.signature hack
- Add save_images() shared utility to converter/__init__.py
- Implement WordConverter.convert() in word_converter.py (moved from mixin.py)
- Implement LatexConverter.convert() in latex_converter.py (moved from mixin.py)
- Slim WordMixin.save_to_word() and LatexMixin.save_to_latex() to ~15 lines each
- Remove inspect.signature runtime reflection from mixin.py (-370 lines net)
- Fix bug: image/chart/seal blocks now use continue instead of returning str
- Fix bug: reference block in LatexConverter returns str, not tuple
- Adjust pp_doctranslation/result.py: rename _to_word(save_path)/_to_latex(save_path)
  to save_to_word()/save_to_latex(), add stub _to_word()/_to_latex() methods
